requirements.txt
pyyaml==6.0.1
requests==2.19.1